Further divide modules
