+++ golem-worker-service-base/src/expression/text.rs
@@ -241,7 +241,8 @@ mod tests {
     fn test_sequence_of_sequence_singleton() {
         let expr_string = "${[[bc]]}";
         let output_expr = from_string(expr_string).unwrap();
-        let expected_expr = Expr::Sequence(vec![Expr::Sequence(vec![Expr::Variable("bc".to_string())])]);
+        let expected_expr =
+            Expr::Sequence(vec![Expr::Sequence(vec![Expr::Variable("bc".to_string())])]);
         assert_eq!(output_expr, expected_expr);
     }
 
@@ -260,7 +261,8 @@ mod tests {
     fn test_sequence_of_tuple_singleton() {
         let expr_string = "${[(bc)]}";
         let output_expr = from_string(expr_string).unwrap();
-        let expected_expr = Expr::Sequence(vec![Expr::Tuple(vec![Expr::Variable("bc".to_string())])]);
+        let expected_expr =
+            Expr::Sequence(vec![Expr::Tuple(vec![Expr::Variable("bc".to_string())])]);
         assert_eq!(output_expr, expected_expr);
     }
