+++ golem-worker-service-base/src/parser/mod.rs
@@ -1,9 +1,12 @@
 use std::fmt;
 
-pub(crate) mod expr_parser;
+pub(crate) use expr::expr_parser;
+
 pub(crate) mod literal_parser;
 pub(crate) mod path_pattern_parser;
 pub(crate) mod place_holder_parser;
+
+mod expr;
 
 pub trait GolemParser<T> {
     fn parse(&self, str: &str) -> Result<T, ParseError>;
